requirements.txt
openai>=1.30.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0
redis>=5.0.0
python-dotenv>=1.0.0
aiogram>=3.10.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
jinja2>=3.1.0
celery[redis]>=5.4.0
